python_backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-core
pydantic
tavily-python
python-dotenv
asyncpg
sqlalchemy
aiosqlite
langchain-anthropic
langchain-google-genai
langchain-groq
langchain-fireworks
langgraph-checkpoint-sqlite
langgraph-checkpoint-postgres
rich
